+++ examples/04_load‎_and_optimize_package‎/optimize.py
@@ -47,28 +47,41 @@ def parse_args():
     return parser.parse_args()
 
 
-# Re-run the conversions and profiling from existing package
 def main():
+    """Load package, optimize and save new package"""
     args = parse_args()
 
     logger.info(f"Loading package from {args.input_path}.")
+    """
+    Load an existing package from the given input path.
+
+    The load function automatically detects the used framework and package version
+    based on the contents of the package."
+    """
     package = nav.package.load(
         path=args.input_path,
     )
 
     logger.info("Optimize package.")
+    """
+    Re-run the conversions, correctmness tests and profiling from existing package.
+    Optimization creates new package.
+    `package` API together with `save`, `load' and `optimize` methods helps with transfering models.
+    It is possible to save the package, load it on different machine and run the optimization again.
+    """
     package = nav.package.optimize(
         package=package,
-        verbose=args.verbose,
-        debug=args.debug,
-        defaults=(not args.no_defaults),
+        verbose=args.verbose,  # enable verbose logs
+        debug=args.debug,  # enable debug logs
+        defaults=(not args.no_defaults),  # use default optimization settings or settings from the package
     )
 
     logger.info(f"Save package to {args.output_path}.")
+    """Save new package to the given output path."""
     nav.package.save(
         package=package,
         path=args.output_path,
-        override=True,
+        override=True,  # override the existing package if it exists.
     )
 
     logger.info("Package saved.")