+++ examples/05_optimize_and_serve_model_on_pytriton/client.py
@@ -11,21 +11,30 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Client for mlp_random example."""
+"""Example client for linear model"""
 import logging
 
-import tensorflow as tf  # pytype: disable=import-error
+import torch  # pytype: disable=import-error
 from pytriton.client import ModelClient  # pytype: disable=import-error
 
 logger = logging.getLogger("client")
 logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(levelname)s - %(name)s: %(message)s")
 
-batch_size = 2
-image_size = (224, 224, 3)
-images_batch = tf.random.uniform(shape=(batch_size,) + image_size, minval=0, maxval=1, dtype=tf.dtypes.float32).numpy()
 
-with ModelClient("localhost", "mlp") as client:
-    logger.info("Sending request")
-    result_dict = client.infer_batch(input__0=images_batch)
+def main():
+    """Prepare data and send inference request to PyTriton server"""
 
-logger.info(f"results: {result_dict}")
+    batch_size = 10
+    data_size = (5,)
+    data_batch = torch.randn(size=(batch_size,) + data_size, dtype=torch.float32).numpy()
+
+    """Use PyTriton client to send inference request."""
+    with ModelClient("localhost", "linear") as client:
+        logger.info("Sending request")
+        result_dict = client.infer_batch(input__0=data_batch)
+
+    logger.info(f"results: {result_dict}")
+
+
+if __name__ == "__main__":
+    main()